Fix #56 Require Java 8+ instead of Java 11+ at runtime
