+++ common/pom.xml
@@ -25,51 +25,16 @@
         <version>0.1.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>mvnd-daemon</artifactId>
+    <artifactId>mvnd-common</artifactId>
 
     <packaging>jar</packaging>
-    <name>Maven Daemon</name>
-
-    <properties>
-        <os.arch>${os.arch}</os.arch><!-- provisio plugin does not see system properties -->
-    </properties>
+    <name>Maven Daemon - Common</name>
 
     <dependencies>
-        <dependency>
-            <groupId>org.jboss.fuse.mvnd</groupId>
-            <artifactId>mvnd-common</artifactId>
-            <exclusions>
-                <exclusion>
-                    <groupId>org.slf4j</groupId>
-                    <artifactId>slf4j-simple</artifactId>
-                </exclusion>
-            </exclusions>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.maven</groupId>
-            <artifactId>maven-embedder</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>io.takari.aether</groupId>
-            <artifactId>takari-local-repository</artifactId>
-        </dependency>
 
-        <!-- Logging -->
         <dependency>
-            <groupId>ch.qos.logback</groupId>
-            <artifactId>logback-classic</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>log4j-over-slf4j</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>jcl-over-slf4j</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>jul-to-slf4j</artifactId>
+            <groupId>org.apache.commons</groupId>
+            <artifactId>commons-compress</artifactId>
         </dependency>
 
         <dependency>
@@ -77,8 +42,13 @@
             <artifactId>jline-terminal</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.codehaus.groovy</groupId>
-            <artifactId>groovy</artifactId>
+            <groupId>org.jline</groupId>
+            <artifactId>jline-terminal-jansi</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-simple</artifactId>
         </dependency>
 
         <dependency>
@@ -105,22 +75,6 @@
                             <goal>sisu-index</goal>
                         </goals>
                         <phase>process-classes</phase>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
-                <groupId>io.takari.maven.plugins</groupId>
-                <artifactId>provisio-maven-plugin</artifactId>
-                <executions>
-                    <execution>
-                        <id>maven-distro</id>
-                        <phase>package</phase>
-                        <goals>
-                            <goal>provision</goal>
-                        </goals>
-                        <configuration>
-                            <outputDirectory>${project.build.directory}/maven-distro</outputDirectory>
-                        </configuration>
                     </execution>
                 </executions>
             </plugin>
