+++ common/src/main/resources/org/jboss/fuse/mvnd/common/build.properties
