+++ common/src/main/resources/org/jboss/fuse/mvnd/common/mvnd.properties.template
