+++ integration-tests/pom.xml
@@ -30,6 +30,8 @@
     <name>Maven Daemon - Integration Tests</name>
 
     <properties>
+        <maven.compiler.target>11</maven.compiler.target>
+        <maven.compiler.source>11</maven.compiler.source>
         <mvnd.home>${project.basedir}/../daemon/target/maven-distro/mvnd-${project.version}-${os.detected.name}-${os.arch}</mvnd.home>
     </properties>
 
@@ -47,6 +49,16 @@
         <dependency>
             <groupId>org.jboss.fuse.mvnd</groupId>
             <artifactId>mvnd-daemon</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.jboss.fuse.mvnd</groupId>
+            <artifactId>mvnd-common</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.jboss.fuse.mvnd</groupId>
+            <artifactId>mvnd-client</artifactId>
             <scope>test</scope>
         </dependency>
         <dependency>
