+++ integration-tests/src/test/java/org/jboss/fuse/mvnd/junit/MvndTestExtension.java
@@ -25,13 +25,13 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Stream;
-import org.jboss.fuse.mvnd.client.BuildProperties;
 import org.jboss.fuse.mvnd.client.Client;
-import org.jboss.fuse.mvnd.client.DaemonInfo;
-import org.jboss.fuse.mvnd.client.DaemonRegistry;
 import org.jboss.fuse.mvnd.client.DefaultClient;
-import org.jboss.fuse.mvnd.client.Environment;
-import org.jboss.fuse.mvnd.client.Layout;
+import org.jboss.fuse.mvnd.common.BuildProperties;
+import org.jboss.fuse.mvnd.common.DaemonInfo;
+import org.jboss.fuse.mvnd.common.DaemonRegistry;
+import org.jboss.fuse.mvnd.common.Environment;
+import org.jboss.fuse.mvnd.common.Layout;
 import org.jboss.fuse.mvnd.jpm.ProcessImpl;
 import org.junit.jupiter.api.extension.AfterAllCallback;
 import org.junit.jupiter.api.extension.BeforeAllCallback;
